Refactored classes into more packages
Now has command, exception, parser, storage, task and ui package
