+++ src/main/java/duke/command/Command.java
@@ -1,4 +1,8 @@
-package duke;
+package duke.command;
+
+import duke.storage.Storage;
+import duke.task.TaskList;
+import duke.ui.Ui;
 
 /**
  * Abstract class Command that represents commands on tasks.
