+++ src/main/java/duke/parser/Parser.java
@@ -1,7 +1,16 @@
-package duke;
+package duke.parser;
 
 import java.util.ArrayList;
 import java.util.Arrays;
+
+import duke.command.AddCommand;
+import duke.command.Command;
+import duke.command.DeleteCommand;
+import duke.command.ExitCommand;
+import duke.command.ListCommand;
+import duke.command.MarkCommand;
+import duke.command.UnmarkCommand;
+import duke.exception.DukeException;
 
 /**
  * Parser class that deals with making sense of the user command.
